requirements.txt
python-telegram-bot==20.8
playwright==1.44.0
playwright-stealth==1.0.6
PyYAML==6.0.1
aiosqlite==0.20.0
httpx==0.27.0
Pillow==10.3.0
pytest==8.0.0
pytest-asyncio==0.23.0
